Updated and Commented Maze Package Objects
I've done as much as I can with the basic objects in the Maze class in terms of commenting and constructing. I've added in a class called Coordinate which stores a row and a col, but if we all deem this to be useless or tedious to work with we can easily remove this and just stick with row and col as individual ints. I imagine the Player class will have methods like "add" and "remove" for the inventory, and I've added in a simple way to count all the Treasures in a stage using static variables and methods. The Goal and Hint classes have been added too.

I reverted my changes to fix a typo in my message, I didn't realise this would also just clear literally everything I had done! Fortunately, I could recover all the work I had done.