Added TileOn parameter to Entities
It became very obvious that Entities should have a link back to the tile that they exist on, if any, so when it comes to removing an Entity from a tile on picking up or updating this it's much easier, as long as we don't end up in a catch-22 situation. Some of the game logic for objects such as the Key, Door, and Treasure have been mostly coded too.