Several File Changes to Fit the revised layout
Door is now an abstract class that KeyDoor and TreasureDoor extend from. TileTypes determine whether a Tile is a floor, wall, hint or goal (and possibly others in the future). Hint is now a HintTile extending from Tile, as it is essentially just a tile with a String appended to it. Player now extends Moveable, a class Teagan created, which will come with a canWalkOn method. Entity is now an interface, as it being an abstract class no longer makes sense. Many of these suggestions were raised to attention by Teagan (TileTypes, Entity as interface, Moveable) so just because I have coded them I would like to give her credit for the idea.