Fixing Player Coordinate when switching levels
Levels are now switching in order without bugs, and the Player is in the right position! Passing the GamePanel as a parameter and updating the Render's Maze ensures that it updates properly, when it wasn't before. I want to note that the code in the nextLevel method is not mine, it is Teagan Stewart's code. I just moved her code out of the Maze class, where it originally was, and into the Game class instead, as the Maze object needed to be overwitten.