Fixing Errors with the Merge
Some of the old changes were not cleared properly and there were some weird artefacts left behind, such as a method in Render severed in two. The only code I wasn't sure how to solve was a line in the Maze class for reloading, so I have simply commented this out.