Adding Hint on the Sidebar Panel
The code is a little unfavourable (needed to use HTML tags to clear the display), but hints are fully implemented. As long as they aren't too long, they will fit inside the display space I have given them. This may only be temporary, or we could use this in the final version of the game.