Added in Swing Timer for Game Loop/Timer Display
To update the GUI every second I have added in a static timer that repeats itself every 500 seconds. I have added in a delay using a boolean so it can be used to display the time on the info panel (which I have implemented a basic numerical timer in, but we have number graphics for visual flair). The main purpose of this is to be used for the game loop, where the moveable characters can be updated.