Basic Replaying Working
Upon level completion, a replay of the level will run. You can use S to Skip the replay, or D to reshow it from any point. Prompts on screen for controls on how to use this, as well as a bar for showing how far through the replay and the current move count will be implemented. Some of the code was repurposed from the Level class, when using System.currentTimeMillis().