Big Updates to Recnplay, now uses JSon
Prior to this update, Recnplay was entirely internally based and so any recordings would not be saved outside of the program. Several things have been added, including a specific timer for replays now, and various classes within Persistence, Application and Recnplay have been updated to make this work. There are still more updates that need to be pulled in, like stopping the timer on the window closing (as the only way to stop an active timer currently is via the console. A recnplay specific panel is ideally going to be added too.