Bigger Updates to RecnPlay Display in the Sidebar
Felt like staying up and coding, so here we go. Recnplay Info Panel is a lot more catered to the replay, which makes sense. It shows the clear time and the current board event of the replay, though more is to be added. Still need to add in key inputs for controlling the replay, for example. As well as this, I would like an update to the json of player where the player's direction is stored, as it would make life far easier here (making Direction.getFacing(..) redundant).