Fixing some Timer related Bugs and More
Fixes a bug in particular where the program waits endlessly on a fram .and wont skip unless instructed to. This is because with the rounding method in place, the game can sometimes have two mazes saved at the exact same time interval. If this is true, it should skip the latter one.