Implemented SaveLevel(int) for exiting without saving.
This method should be called when the user has chosen to exit the game WITHOUT saving. It saves the number of the level the user was on, so that when resumeGame is called, it starts the game from a fresh copy of that level number, not a random previous save.