THIS IS NOT MY CODE. THIS IS BECAUSE A MERGE BROKE LAST MINUTE
Even though it says I have added all this code, believe me I did not code this entire project. There is more than enough evidence throughout the prior commits to prove that. Though, I did add in some of the toStrings() to some of the Entities that did not have them.